backend/requirements.txt
# requirements.txt
Flask==2.3.3
flask-cors==4.0.0
opencv-python==4.8.1.78
pytesseract==0.3.10
Pillow==10.0.1
numpy==1.24.3

# Database requirements
psycopg2-binary==2.9.7
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5

# Authentication requirements
Flask-JWT-Extended==4.5.3
Flask-Bcrypt==1.0.1
email-validator==2.1.0.post1

# Development utilities
python-dotenv==1.0.0

# Production server
gunicorn==20.1.0